section_4/labs/040-080-containerize-our-model/requirements.txt
pillow==10.4.0
gunicorn==23.0.0
Flask==3.1.0
